requirements.txt
aiogram==2.25.1
python-dotenv==1.0.0
openai==0.28
SpeechRecognition==3.10.0
pydub==0.25.1
